refactor: circleMember 패키지 분리
