--- src/main/java/com/circle/circle_backend/circle/domain/enums/UserRole.java
@@ -1,7 +1,0 @@
-package com.circle.circle_backend.circle.domain.enums;
-
-public enum UserRole {
-
-    MEMBER,
-    ADMIN
-}
