+++ src/main/java/com/circle/circle_backend/circleMember/domain/enums/UserRole.java
@@ -1,0 +1,7 @@
+package com.circle.circle_backend.circleMember.domain.enums;
+
+public enum UserRole {
+
+    MEMBER,
+    ADMIN
+}
