+++ src/main/java/com/circle/circle_backend/circleMember/infrastructure/CircleMemberJpaRepository.java
@@ -1,6 +1,6 @@
-package com.circle.circle_backend.circle.infrastructure;
+package com.circle.circle_backend.circleMember.infrastructure;
 
-import com.circle.circle_backend.circle.infrastructure.entity.CircleMemberEntity;
+import com.circle.circle_backend.circleMember.infrastructure.entity.CircleMemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
