+++ src/main/java/com/circle/circle_backend/circleMember/service/CircleMemberRepository.java
@@ -1,7 +1,8 @@
-package com.circle.circle_backend.circle.service;
+package com.circle.circle_backend.circleMember.service;
 
-import com.circle.circle_backend.circle.domain.CircleMember;
-import com.circle.circle_backend.circle.infrastructure.entity.CircleMemberEntity;
+
+import com.circle.circle_backend.circleMember.domain.CircleMember;
+import com.circle.circle_backend.circleMember.infrastructure.entity.CircleMemberEntity;
 
 import java.util.List;
 import java.util.Optional;
